fix bugs #2

+++ book-info/src/main/java/dl/iitu/bookinfo/Book/Repository.java
@@ -2,6 +2,6 @@ package dl.iitu.bookinfo.Book;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Repository extends JpaRepository<Book, Integer>{
+public interface Repository extends JpaRepository<Book, Long>{
 
 }
