+++ user-info/src/main/java/dl/iitu/userinfo/Repository.java
@@ -1,7 +1,6 @@
 package dl.iitu.userinfo;
 
-import dl.iitu.userinfo.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Repository extends JpaRepository<User, Integer> {
+public interface Repository extends JpaRepository<User, Long> {
 }
